allow compiling the whole LDraw library to an OpenSCAD library
This restructures the code to be more flexible and allows three modes of
operation:
1. Compile a single model file into a single self-contained OpenSCAD
   file. This is the old behavior.
2. Compile a single model file into an OpenSCAD file that relies on the
   presence of an LDraw OpenSCAD library. This is the new default.
3. Compile the whole LDraw library into an OpenSCAD library to support
   files translated in operation mode 2.

Technically there is also the option to translate the whole LDraw
library into a single library file. Due to the size of the LDraw
library, this has no practical use though since this causes OpenSCAD to
eat away all our memory when attempting to parse the result.

Additionally, this comes with the following features and benefits:
- Make the transfer of the original file contents as comments switchable
  with having them included being the default.
- New way of handling MPD files that seems more robust than the previous
  one. In particular, this can successfully translate the files from the
  model library on the LDraw site.