add BFC support
This adds basic BFC support.  We can't properly support INVERTNEXT
with the current implementation though.  One solution to fill this
gap would be to change the implementation to a purely functional
definition for the subparts, which would kill the color support on
them though.  Maybe we find a solution for that one later.