move quite some logic from Python into OpenSCAD
This changes the implementation to make more use of the functional
power or OpenSCAD, leaving the translated lines more similar to the
original LDraw files.  As a positive side effect this fixes two
issues with the previous implementation:
1. Colours now only override color index 16 instead of any colour as
   specified in the LDraw file specification.
2. BWC inversion now properly works.
In addition to those benefits we can now start implementing OpenSCAD
functions calculating bounding boxes or other interesting parameters
about the LDraw model.