allow translation of complete model directories
When the input parameter is a directory we translate all models in this
directory and all of its subdirectories.

Also fix a bug where internal references in MPD files were assumed to be
case sensitive where official models expect them to be case insensitive
apparently.

We also make the location of the LDraw library, the location of the
OpenSCAD libraries, and the name of the generated OpenSCAD library an
option instead of hard-coding their locations or name in the code. The
values default to their previous values with the exception of the
OpenSCAD library name defaulting to 'LDraw' since 'openscad' was not a
really brilliant name choice for an OpenSCAD library.